Implement limiting map drawing to a circular region
A single class, VisibilityLimit, controls what's visible. Right now this class is hardcoded to use a single circle as the visibility limit. In the future, I guess in the future it could store a list of shapes, and then query each of them whether that part of the map should be visible.

Care has been taken to make sure that regions and chunks completely outside the visibility limit are ignored. Therefore, these chunks aren't loaded when using /pl3xmap fullrender (or radiusrender, or background rendering).